[v0.3.0]
[v0.3.0]
- [update] Updated QMapControl to version 0.9.7.8.
- [update] Updated QCustomPlot to version 1.3.0.
- [update] Updated copyright to "2014 - 2015".
- [new] Track on the map is highlighted when the mouse cursor is moved
over it.
- [new] Show some point properties when the mouse cursor is over the
point.
- [new] Added functionality to insert/delete track points.
- [new] Added functionality to set the start time of a track.
- [new] Added functionality to move a single track point using SHIFT +
mouse click.
- [new] Added a calendar to select a track by date.
- [new] Diagram curves can be changed to show other properties.
- [new] Support SpoQ files (.act and .xml).
- [new] Support Garmin's TrackPointExtension heart rate extension.
- [fix] Ignore upper/lower case to detect file type.
- [fix] Draw tracks  with 0 or only 1 point correctly.
- [fix] Changed way of iteration in GPX_wrapper.
- [fix] Fixed bug in NMEA sentence parser.

+++ GPXLab/dialogs/dialog_modelproperties.cpp
@@ -1,5 +1,5 @@
 /****************************************************************************
- *   Copyright (c) 2014 Frederic Bourgeois <bourgeoislab@gmail.com>         *
+ *   Copyright (c) 2014 - 2015 Frederic Bourgeois <bourgeoislab@gmail.com>  *
  *                                                                          *
  *   This program is free software: you can redistribute it and/or modify   *
  *   it under the terms of the GNU General Public License as published by   *
@@ -107,7 +107,7 @@ void Dialog_modelProperties::on_Dialog_modelProperties_accepted()
         // Tab: Author
         metadata.author.name = ui->lineEditAuthorName->text().toStdString();
         string email = ui->lineEditAuthorEmail->text().toStdString();
-        int atPos = email.find_last_of("@");
+        size_t atPos = email.find_last_of("@");
         if (atPos > 0)
         {
             metadata.author.email.id = email.substr(0, atPos);
